Split facebook module into a package.
